move common functions into utils
