+++ gen3-client/g3cmd/utils.go
@@ -8,9 +8,14 @@ import (
 	"log"
 	"net/http"
 	"os"
+	"path"
+	"path/filepath"
+	"strconv"
 	"strings"
+	"sync"
 	"time"
 
+	"github.com/uc-cdis/gen3-client/gen3-client/commonUtils"
 	"github.com/uc-cdis/gen3-client/gen3-client/logs"
 
 	"github.com/uc-cdis/gen3-client/gen3-client/jwt"
@@ -101,11 +106,11 @@ func GenerateUploadRequest(furObject FileUploadRequestObject, file *os.File) (Fi
 		writer = io.MultiWriter(pw, bar)
 		if _, err = io.Copy(writer, file); err != nil {
 			logs.WriteToFailedLog(true)
-			log.Fatalf("io.Copy error: %s\n", err)
+			log.Printf("io.Copy error: %s\n", err)
 		}
 		if err = pw.Close(); err != nil {
 			logs.WriteToFailedLog(true)
-			log.Fatalf("Pipe writer close error: %s\n", err)
+			log.Printf("Pipe writer close error: %s\n", err)
 		}
 	}()
 
@@ -116,4 +121,213 @@ func GenerateUploadRequest(furObject FileUploadRequestObject, file *os.File) (Fi
 	furObject.Bar = bar
 
 	return furObject, err
+}
+
+func validateFilePath(filePaths []string) []string {
+	validatedFilePaths := make([]string, 0)
+	for _, filePath := range filePaths {
+		if _, err := os.Stat(filePath); os.IsNotExist(err) {
+			log.Printf("The file you specified \"%s\" does not exist locally", filePath)
+			continue
+		}
+
+		file, err := os.Open(filePath)
+		if err != nil {
+			log.Printf("File open error")
+			continue
+		}
+
+		if fi, _ := file.Stat(); fi.IsDir() {
+			continue
+		}
+
+		if logs.ExistsInSucceededLog(filePath) {
+			fmt.Println("File \"" + filePath + "\" has been found in local submission history and has be skipped for preventing duplicated submissions.")
+			continue
+		} else {
+			logs.AddToFailedLogMap(filePath, "", true)
+		}
+		validatedFilePaths = append(validatedFilePaths, filePath)
+		file.Close()
+	}
+	logs.WriteToFailedLog(false)
+	return validatedFilePaths
+}
+
+func processFilename(filePath string, includeSubDirName bool) (fileInfo, error) {
+	var err error
+	filename := filepath.Base(filePath)
+	if includeSubDirName {
+		presentDirname := strings.TrimSuffix(commonUtils.ParseRootPath(uploadPath), commonUtils.PathSeparator+"*")
+		subFilename := strings.TrimPrefix(filePath, presentDirname)
+		dir, _ := filepath.Split(subFilename)
+		if dir != "" && dir != commonUtils.PathSeparator {
+			filename = strings.TrimPrefix(subFilename, commonUtils.PathSeparator)
+			filename = strings.Replace(filename, commonUtils.PathSeparator, ".", -1)
+		} else {
+			err = errors.New("Include subdirectory names will only works if the file is under at least one subdirectory.")
+		}
+	}
+	return fileInfo{filePath, filename}, err
+}
+
+func getFullFilePath(filePath string, filename string) (string, error) {
+	fi, err := os.Stat(filePath)
+	if err != nil {
+		fmt.Println(err)
+		return "", err
+	}
+	switch mode := fi.Mode(); {
+	case mode.IsDir():
+		if strings.HasSuffix(filePath, "/") {
+			return filePath + filename, nil
+		}
+		return filePath + "/" + filename, nil
+	case mode.IsRegular():
+		return "", errors.New("in manifest upload mode filePath must be a dir")
+	default:
+		return "", errors.New("full file path creation unsuccessful")
+	}
+}
+
+func validateObject(objects []ManifestObject) []FileUploadRequestObject {
+	furObjects := make([]FileUploadRequestObject, 0)
+	for _, object := range objects {
+		guid := object.ObjectID
+		// Here we are assuming the local filename will be the same as GUID
+		filePath, err := getFullFilePath(uploadPath, object.ObjectID)
+		if err != nil {
+			log.Println(err.Error())
+			continue
+		}
+
+		if _, err := os.Stat(filePath); os.IsNotExist(err) {
+			log.Println("The file you specified \"%s\" does not exist locally.", filePath)
+			continue
+		}
+
+		furObject := FileUploadRequestObject{FilePath: filePath, FileName: path.Base(filePath), GUID: guid}
+		furObjects = append(furObjects, furObject)
+	}
+	return furObjects
+}
+
+func uploadFile(furObject FileUploadRequestObject) error {
+	fmt.Println("Uploading data ...")
+	furObject.Bar.Start()
+
+	client := &http.Client{}
+	resp, err := client.Do(furObject.Request)
+	if err != nil {
+		logs.AddToFailedLogMap(furObject.FilePath, furObject.PresignedURL, false)
+		furObject.Bar.Finish()
+		return errors.New("Error occurred during upload: " + err.Error())
+	}
+	if resp.StatusCode != 200 {
+		logs.AddToFailedLogMap(furObject.FilePath, furObject.PresignedURL, false)
+		furObject.Bar.Finish()
+		return errors.New("Upload request got a non-200 response with status code " + strconv.Itoa(resp.StatusCode))
+	}
+	furObject.Bar.Finish()
+	fmt.Printf("Successfully uploaded file \"%s\" to GUID %s.\n", furObject.FilePath, furObject.GUID)
+	logs.DeleteFromFailedLogMap(furObject.FilePath, true)
+	logs.WriteToSucceededLog(furObject.FilePath, furObject.GUID, false)
+	return nil
+}
+
+func initBathUploadChannels(numParallel int, inputSliceLen int) (int, chan *http.Response, chan error, []FileUploadRequestObject) {
+	workers := numParallel
+	if workers < 1 || workers > inputSliceLen {
+		workers = inputSliceLen
+	}
+	respCh := make(chan *http.Response, inputSliceLen)
+	errCh := make(chan error, inputSliceLen)
+	batchFURSlice := make([]FileUploadRequestObject, 0)
+	return workers, respCh, errCh, batchFURSlice
+}
+
+func batchUpload(furObjects []FileUploadRequestObject, workers int, respCh chan *http.Response, errCh chan error) {
+	bars := make([]*pb.ProgressBar, 0)
+	respURL := ""
+	var err error
+	var guid string
+	var filename string
+
+	for i := range furObjects {
+		if furObjects[i].GUID == "" {
+			respURL, guid, filename, err = GeneratePresignedURL(furObjects[i].FilePath, includeSubDirName)
+			if err != nil {
+				logs.AddToFailedLogMap(furObjects[i].FilePath, respURL, false)
+				errCh <- err
+				continue
+			}
+			furObjects[i].PresignedURL = respURL
+			furObjects[i].GUID = guid
+			furObjects[i].FileName = filename
+		}
+		file, err := os.Open(furObjects[i].FilePath)
+		if err != nil {
+			logs.AddToFailedLogMap(furObjects[i].FilePath, furObjects[i].PresignedURL, false)
+			errCh <- errors.New("File open error: " + err.Error())
+			continue
+		}
+		defer file.Close()
+
+		furObjects[i], err = GenerateUploadRequest(furObjects[i], file)
+		if err != nil {
+			file.Close()
+			logs.AddToFailedLogMap(furObjects[i].FilePath, furObjects[i].PresignedURL, false)
+			errCh <- errors.New("Error occurred during request generation: " + err.Error())
+			continue
+		}
+		bars = append(bars, furObjects[i].Bar)
+	}
+
+	furObjectCh := make(chan FileUploadRequestObject, len(furObjects))
+
+	pool, err := pb.StartPool(bars...)
+	if err != nil {
+		for _, furObject := range furObjects {
+			logs.AddToFailedLogMap(furObject.FilePath, furObject.PresignedURL, false)
+		}
+		errCh <- errors.New("Error occurred during starting progress bar pool: " + err.Error())
+		return
+	}
+
+	client := &http.Client{}
+	wg := sync.WaitGroup{}
+	for i := 0; i < workers; i++ {
+		wg.Add(1)
+		go func() {
+			for furObject := range furObjectCh {
+				if furObject.Request != nil {
+					resp, err := client.Do(furObject.Request)
+					if err != nil {
+						logs.AddToFailedLogMap(furObject.FilePath, furObject.PresignedURL, true)
+						errCh <- err
+					} else {
+						if resp.StatusCode != 200 {
+							logs.AddToFailedLogMap(furObject.FilePath, furObject.PresignedURL, true)
+						} else { // Succeeded
+							respCh <- resp
+							logs.DeleteFromFailedLogMap(furObject.FilePath, true)
+							logs.WriteToSucceededLog(furObject.FilePath, furObject.GUID, true)
+							logs.ScoreBoard[0]++
+						}
+					}
+				} else if furObject.FilePath != "" {
+					logs.AddToFailedLogMap(furObject.FilePath, furObject.PresignedURL, true)
+				}
+			}
+			wg.Done()
+		}()
+	}
+
+	for i := range furObjects {
+		furObjectCh <- furObjects[i]
+	}
+	close(furObjectCh)
+
+	wg.Wait()
+	pool.Stop()
 }
